Make sure multi-step touch actions work
