+++ lib/commands/touch.js
@@ -148,17 +148,32 @@ commands.performTouch = async function (gestures) {
     // some things are special
     return await this.doTouchDrag(gestures);
   } else {
-    // `press` without a wait is too slow and gets interpretted as a `longPress`
-    if (actions[actions.length - 2] === 'press' && actions[actions.length - 1] === 'release') {
-      actions[actions.length - 2] = 'tap';
-      gestures[gestures.length - 2].action = 'tap';
-    }
+    if (actions.length === 2) {
+      // `press` without a wait is too slow and gets interpretted as a `longPress`
+      if (_.first(actions) === 'press' && _.last(actions) === 'release') {
+        actions[0] = 'tap';
+        gestures[0].action = 'tap';
+      }
 
-    // the `longPress` and `tap` methods release on their own
-    if ((actions[actions.length - 2] === 'tap' ||
-      actions[actions.length - 2] === 'longPress') && actions[actions.length - 1] === 'release') {
-      gestures.pop();
-      actions.pop();
+      // the `longPress` and `tap` methods release on their own
+      if ((_.first(actions) === 'tap' || _.first(actions) === 'longPress') && _.last(actions) === 'release') {
+        gestures.pop();
+        actions.pop();
+      }
+    } else {
+      // longpress followed by anything other than release should become a press and wait
+      if (actions[0] === 'longPress') {
+        actions = ['press', 'wait', ...actions];
+
+        let press = gestures.shift();
+        press.action = 'press';
+        let wait = {
+          action: 'wait',
+          options: {ms: press.options.duration || 1000}
+        };
+        delete press.options.duration;
+        gestures = [press, wait, ...gestures];
+      }
     }
 
     // fix release action then perform all actions
